Fix selected text from iframe
Fixed text selection from iframes in plugins "insipio-tts",
"espeak-tts", "dictionary" and "dictionary-stemmer". However the iframe
must be within the same domain due to browser security restrictions. We
tried to get the text from the clipboard in the case that the iframe is
from other domain, but it's working just for I.E at the moment.